Revert "RFTR: UAM/UAV Components Addition:recycle:. Closes #137 (#144)"
This reverts commit 9e66df5db8041dcd464b6e1885b585657b545af5.
